Adds a bunch of stuff having to do with API authentication and Keycloak integration.
- Adds a container for APISIX, which sits in front of the nginx instance
- Adds support for storing OAuth tokens for the Keycloak admin API
- Adds a couple of middlewares to handle authentication when routing API requests through APISIX, and one for the same when routing through Traefik with a ForwardAuth that then runs through Keycloak
- Adjusted the config directory structure

Note that there is no Traefik setup here - you will need to do that on your own. Ask me if you want the config for that, or maybe I'll throw it in a future commit.

The APISIX instance is configured to run but has no routes defined out of the box - so it won't work for anything until you do that.

The nginx server still exposes port 8073 so you can still get to it that way. The auth middleware is set up so that it'll fall back to the normal Model auth so you can log in as per usual if there's no gateway in between you and the API.

This is not meant to be merged into main - this is instead the culmination of two experiments that I wanted to get thrown into something tracked so that I can cherry pick the winner out of to put into main.