Update the session invalidation code
This was using _remove_invalid_user from the upstream RemoteUserBackend middleware - but that only works if the middleware is specifically RemoteUserBackend, so now instead log them out using auth.logout.

I am not real sure how I managed to hit this so this is hard to test, but I did manage to get into a different user's session.