Some updates to shore this up with upstream
This removes the apisix specific Compose file, so needed to pull the Keycloak stuff into the main one. (This is fine - APISIX is now a hard dependency anyway.) Also, added env settings to example file and updated the docs a bit.

This introduces a change: you must now supply a keystore password for Keycloak. See docs. This was specified in the Compose file before.