Updates to pull the APISIX config into the config.yaml file
This will allow devs to bootstrap the system without having to hit the API manually to define resources in APISIX. Plus, it gets a baseline set up for the config that we can maybe use elsewhere (like in API7 gateways).

This also removes the home script below - this shouldn't really have ever been committed so not sure how it did.