More stuff to shore up OpenAPI generation
Adds some extend_schemas/etc to places to help the API generation along, adding more type hints and stuff. Also simplified the IsAdminUserOrReadOnly permission class.

This won't pass generator specs - there are a bunch of warnings that get generated which causes it to fail. These need to be fixed but fixing these really steps on the toes of another issue that's out for the basket/checkout APIs, so punting to that.