Reworks route plugins
- Move _most_ API routes _out_ of OIDC protection - you'll need a Django session, it just won't start one for you for most routes now
- Reassign `cart`, `admin`, and `establish_session` into OIDC control since these get hit by the end user directly
- Fix an error in the URL generation for `establish_session`
- Update the docker-compose to tag latest for the apisix container

Prior to this change, the docker-compose env forced APISIX to run as amd64 - there was a reason for this which has been lost to time. In testing, though, the arm64 version works fine too, so this removes the platform tag. **On Apple Silicon** (or other arm64) you'll have to manually pull the apisix image or Docker likely won't figure it out and will continue to try to use the amd64 one, which won't work.