Adds a "create_with_products" API for baskets.
This allows you to create a basket that has multiple products in it. You can optionally add a discount code in too. Products all have to be from the same system, as you can only specify the system slug once. SKUs that aren't found error out the entire request. Discount codes that don't should emit an error but won't cancel the transaction.