Fixes for tests - had an issue with integrated system name not being unique; changes to force auth to work with headers
The DRF APIClient force_authenticate sets the user in the request but then the APISIX middleware strips that out, because it doesn't match what it's seeing in the headers. (Which is nothing!) So, this adds some code to fire an X-Userinfo header into the APIClient so that it works properly.

This APIClient stuff should move into ol-django apigateway.